Tool cleanup
